urls.py is now compatible with Django 1.10.
This removes a warning that was showing up in Django 1.9. Also, this change
should be compatible (according to the docs) with Django versions going
back to 1.5, but I've only tried it with Django 1.7. Finally, I removed
some pre-Django 1.4 compatibility imports, because I assume such an old
version isn't supported anymore.